Move methods to check WP functions to a trait for re-usability
